sm8550-common: Change odm and vendor partitions to erofs Hopefully this will help with deflicker and a few other camera blobs returning a null pointer reference when building sometimes.
All we have to do when that happens is rebuild without changing anything and they tend to work again which makes me think its partition structure.

Originally all partitions were erofs and my old tree was also erofs and this never happened. So, let's change odm and vendor, where the blobs are to erofs and keep the others as ext4 so we can flash gapps.

Thank you to Kacper for testing this one with me :)

test: dirty flash and check partition file system type, confirm camera works.

Change-Id: If00a052c800c5268477d65c237faf656b3eddb4e

Co-authored-by: lonelystartb9x <lonelystartb9x@gmail.com>
Co-authored-by: TheMysticle <36317711+TheMysticle@users.noreply.github.com>